Fix conformance tests;

+++ Compiler/src/ConstantFolding.h
@@ -31,7 +31,7 @@ struct Constant
         bool valueBoolean;
         double valueNumber;
         float valueVector[4];
-        float valueQuaternion[4];
+        short valueQuaternion[4];
         const char* valueString = nullptr; // length stored in stringLength
     };
